+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/shape/SVGShapeExporter.java
@@ -12,7 +12,8 @@
  * Lesser General Public License for more details.
  * 
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library. */
+ * License along with this library.
+ */
 package com.jpexs.decompiler.flash.exporters.shape;
 
 import com.jpexs.decompiler.flash.SWF;
@@ -200,6 +201,11 @@ public class SVGShapeExporter extends DefaultSVGShapeExporter {
         path.setAttribute("stroke", "url(#gradient" + id + ")");
         path.setAttribute("fill", "none");
         exporter.addToDefs(gradient);
+    }
+
+    @Override
+    public void lineBitmapStyle(int bitmapId, Matrix matrix, boolean repeat, boolean smooth, ColorTransform colorTransform) {
+        //TODO
     }
 
     @Override