+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - AS3 - Remove trait which is outside class
 
+### Fixed
+- Flash viewer - bitmap line style
+
 ## [14.2.1] - 2021-03-13
 ### Added
 - Placeobject display and edit - raw editor on right side